% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{Eda}
\usage{
eda(input_data)
}
\arguments{
\item{input_data}{(dpylyr dataframe) input dataframe to be analyzed- Required}
}
\value{
(list) List that contains a histogram and summary statistics for each column.
}
\description{
Eda
}
\examples{


result <- eda(mtcars)
hist_mpg <- result$histograms[[1]]
stats_mpg <- result$stats$mpg
}
