% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute Missing Data}
\usage{
impute(data)
}
\arguments{
\item{data}{tibble or dataframe including only model features}
}
\value{
tibble with imputed features
}
\description{
Automatically imputes missing data by detecting a features class and
comparing test results of multiple models using different imputation methods
}
\examples{
my_data = tibble(...)
imputed_data = impute(my_data)
}
