% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_summary}
\alias{impute_summary}
\title{Impute Missing Data}
\usage{
impute_summary(df)
}
\arguments{
\item{df}{dataframe containing missing values to impute}
}
\value{
a list object with attributes containing summary statistics and
imputed data of different methods
}
\description{
Consolidates several different imputation methods inlcuding summary statistics
}
