% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{Eda}
\usage{
eda(input_data)
}
\arguments{
\item{input_data}{(dpylyr dataframe) input dataframe to be analyzed- Required}
}
\value{
(list) List that contains a histogram and summary statistics for each column.
}
\description{
Eda
}
\examples{
my_data <- vector(mode="list", length=1)
names(my_data) <- c("tic")
my_data[[1]] <- c(1,2,3,4,5)

result = eda(my_data)
hist_tic = result$histograms$tic
stats_tic = result$stats$tic
}
