% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{Preproc}
\usage{
preproc(X)
}
\arguments{
\item{X}{(dpylyr dataframe) X_train dateframe - Required}
}
\value{
(dataframe) Preprocessed df with scaling or encoding
}
\description{
Preproc
}
\examples{
result = preproc(mtcars)
processed_X = result
}
