% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{Preproc}
\usage{
preproc(X_train, X_test = None)
}
\arguments{
\item{X_train}{(dpylyr dataframe) X_train dateframe - Required}

\item{X_test}{(dpylyr dataframe)  X_test dataframe - Default None}
}
\value{
(tuple) Tuple containing two dataframes (X_train, X_test)
}
\description{
Preproc
}
\examples{
result = preproc(X_train, X_test)
processed_X_train = result[0]
processed_X_test = result[1]
}
